service/serializers.py
from rest_framework import serializers

from .models import Deals


class DealsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Deals
        fields = '__all__'


class ClientsSerializer(serializers.Serializer):
    username = serializers.CharField()
    spent_money = serializers.IntegerField()
    gems = serializers.ListField(child=serializers.CharField())
